pyproject.toml
[project]
name = "rotary-phone-audio-guestbook"
version = "0.1.0"
description = "This project transforms a rotary phone into a voice recorder for use at special events (i.e. wedding audio guestbook, etc.)."
readme = "README.md"
requires-python = ">=3.9.2"
dependencies = [
    "flask>=3.0.3",
    "gevent>=24.11.1",
    "gpiozero>=2.0.1",
    "gunicorn>=23.0.0",
    "ruamel-yaml>=0.18.6",
]
